Move out refocus listener
